Initial version of riak_column
riak_column is a concurrent dict abstraction on top of riak.
It's a variant of vector map, which gracefully handles many
updates because the contents is split into groups controlled
by bit-prefix of the sha of the key.  Intended use is for
secondary indices and/or materialized views on top of riak.